Refactoring usart.c

+++ Code/inc/usart.h
@@ -50,6 +50,6 @@ uint8_t usart1_dma_tx(uint8_t * txData, uint16_t buff_size, uint32_t timeout);
 
 #endif
 
-extern volatile uint32_t usart_buffer_rx[USART_SIZE_BUFFER];
+extern volatile uint8_t usart_buffer_rx[USART_SIZE_BUFFER];
 
 #endif
